pyproject.toml
[project]
name = "rewiring-aotearoa-electrification-progress-tracker"
version = "0.1.0"
description = "Data engineering project for tracking electrification progress in Aotearoa New Zealand"
authors = [
    {name = "Good Data Institute"}
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "requests>=2.32.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "duckdb>=1.1.0",
    "streamlit>=1.39.0",
    "shiny>=1.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
